@@ -14,112 +14,100 @@
   "dependencies": {
     "classnames": "2.2.6",
     "clipboard": "2.0.4",
-    "codemirror": "5.44.0",
-    "color": "3.1.0",
-    "core-js": "2.6.5",
-    "file-saver": "2.0.1",
+    "codemirror": "5.48.2",
+    "color": "3.1.2",
+    "core-js": "3.1.4",
+    "file-saver": "2.0.2",
     "golden-layout": "1.5.9",
     "isomorphic-fetch": "2.2.1",
     "jed": "1.1.1",
-    "jquery": "3.3.1",
+    "jquery": "3.4.1",
     "keymaster": "1.6.2",
     "lz-string": "1.4.4",
-    "markdown-it": "8.4.2",
+    "markdown-it": "9.0.1",
     "node-noop": "1.0.0",
     "po2json": "0.4.5",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
-    "react-redux": "6.0.1",
-    "react-router": "4.3.1",
-    "react-router-dom": "4.3.1",
-    "redux": "4.0.1",
+    "react": "16.8.6",
+    "react-dom": "16.8.6",
+    "react-redux": "7.1.0",
+    "react-router": "5.0.1",
+    "react-router-dom": "5.0.1",
+    "redux": "4.0.4",
     "redux-devtools-extension": "2.13.8",
     "redux-thunk": "2.3.0",
     "reflect-metadata": "0.1.13",
-    "tslib": "1.9.3",
-    "typesafe-actions": "3.2.1"
+    "tslib": "1.10.0",
+    "typesafe-actions": "4.4.2"
   },
   "devDependencies": {
     "@types/chai": "4.1.7",
     "@types/chai-as-promised": "7.1.0",
-    "@types/classnames": "2.2.7",
+    "@types/classnames": "2.2.9",
     "@types/clipboard": "2.0.1",
-    "@types/codemirror": "0.0.72",
+    "@types/codemirror": "0.0.76",
     "@types/color": "3.0.0",
-    "@types/core-js": "2.5.0",
-    "@types/enzyme": "3.9.0",
+    "@types/core-js": "2.5.2",
+    "@types/enzyme": "3.10.3",
     "@types/enzyme-adapter-react-16": "1.0.5",
-    "@types/file-saver": "2.0.0",
+    "@types/file-saver": "2.0.1",
     "@types/history": "4.7.2",
     "@types/isomorphic-fetch": "0.0.35",
-    "@types/jquery": "3.3.29",
-    "@types/jsdom": "12.2.2",
+    "@types/jquery": "3.3.30",
+    "@types/jsdom": "12.2.4",
     "@types/keymaster": "1.6.28",
     "@types/lz-string": "1.3.33",
-    "@types/markdown-it": "0.0.7",
-    "@types/mocha": "5.2.6",
-    "@types/node": "11.9.6",
-    "@types/react": "16.8.6",
-    "@types/react-dom": "16.8.2",
-    "@types/react-redux": "7.0.1",
-    "@types/react-router": "4.4.4",
-    "@types/react-router-dom": "4.3.1",
-    "@types/react-test-renderer": "16.8.1",
-    "@types/redux-mock-store": "1.0.0",
-    "@types/sinon": "7.0.8",
+    "@types/markdown-it": "0.0.8",
+    "@types/mocha": "5.2.7",
+    "@types/node": "12.6.8",
+    "@types/react": "16.8.23",
+    "@types/react-dom": "16.8.5",
+    "@types/react-redux": "7.1.1",
+    "@types/react-router": "5.0.3",
+    "@types/react-router-dom": "4.3.4",
+    "@types/react-test-renderer": "16.8.3",
+    "@types/redux-mock-store": "1.0.1",
+    "@types/sinon": "7.0.13",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
-    "clean-webpack-plugin": "1.0.1",
-    "copy-webpack-plugin": "5.0.0",
-    "css-loader": "2.1.0",
-    "enzyme": "3.9.0",
-    "enzyme-adapter-react-16": "1.10.0",
-    "file-loader": "3.0.1",
-    "fork-ts-checker-webpack-plugin": "0.5.2",
+    "clean-webpack-plugin": "3.0.0",
+    "copy-webpack-plugin": "5.0.4",
+    "css-loader": "3.1.0",
+    "enzyme": "3.10.0",
+    "enzyme-adapter-react-16": "1.14.0",
+    "file-loader": "4.1.0",
+    "fork-ts-checker-webpack-plugin": "1.4.3",
     "git-revision-webpack-plugin": "3.0.3",
     "html-webpack-plugin": "3.2.0",
-    "husky": "1.3.1",
+    "husky": "3.0.2",
     "ignore-styles": "5.0.1",
-    "jsdom": "13.2.0",
+    "jsdom": "15.1.1",
     "json-loader": "0.5.7",
     "less": "3.9.0",
-    "less-loader": "4.1.0",
-    "lint-staged": "8.1.4",
-    "mini-css-extract-plugin": "0.5.0",
-    "mocha": "6.0.2",
+    "less-loader": "5.0.0",
+    "mini-css-extract-plugin": "0.8.0",
+    "mocha": "6.2.0",
     "pofile": "1.0.11",
-    "prettier": "1.16.4",
-    "raw-loader": "1.0.0",
+    "prettier": "1.18.2",
+    "pretty-quick": "1.11.1",
+    "raw-loader": "3.1.0",
     "redux-mock-store": "1.5.3",
-    "sinon": "7.2.5",
+    "sinon": "7.3.2",
     "style-loader": "0.23.1",
-    "ts-loader": "5.3.3",
-    "ts-node": "8.0.2",
+    "ts-loader": "6.0.4",
+    "ts-node": "8.3.0",
     "tsconfig-paths": "3.8.0",
-    "tslint": "5.13.1",
+    "tslint": "5.18.0",
     "tslint-sonarts": "1.9.0",
-    "typescript": "3.3.3333",
-    "url-loader": "1.1.2",
-    "webpack": "4.29.6",
-    "webpack-bundle-analyzer": "3.0.4",
-    "webpack-cli": "3.2.3",
-    "webpack-dev-server": "3.2.1"
+    "typescript": "3.5.3",
+    "url-loader": "2.1.0",
+    "webpack": "4.38.0",
+    "webpack-bundle-analyzer": "3.4.1",
+    "webpack-cli": "3.3.6",
+    "webpack-dev-server": "3.7.2"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "ignore": [
-      "**/*.min.js",
-      "**/*.min.css"
-    ],
-    "linters": {
-      "*.{ts,tsx,js,json,css,less,md}": [
-        "prettier --write",
-        "git add"
-      ]
+      "pre-commit": "pretty-quick --staged"
     }
   }
 }
