@@ -6,4 +6,4 @@ import { GoldenLayoutHelper } from "./golden-layout.helper";
  */
 export const ReactGoldenLayoutHelperContext: React.Context<
   GoldenLayoutHelper | undefined
-> = React.createContext(undefined);
+> = React.createContext(undefined as any);
